Changed database schema, added category support

+++ app/src/main/java/com/callisto/fusion/db/TaskCategory.java
@@ -18,10 +18,10 @@ import android.arch.persistence.room.PrimaryKey;
 public class TaskCategory {
 
     @PrimaryKey(autoGenerate = true)
-    public int TaskCategoryID;
+    public long TaskCategoryID;
 
-    public int taskID;
+    public long taskID;
 
-    public int categoryID;
+    public long categoryID;
 
 }
